Zdev

+++ src/services/Project.ts
@@ -23,24 +23,18 @@ export const list = async (
   req: Request,
   res: Response
 ): Promise<Response | void> => {
-  const projects = await getConnection().getRepository(Project).find();
-  return res.status(OK).json({ projects });
-};
-
-/******************************************************************************
- *                      Get My Projects - "POST /api/projects/myprojects"
- ******************************************************************************/
-export const getmine = async (
-  req: Request,
-  res: Response
-): Promise<Response | void> => {
-  const { id } = req.body;
-  const projects = await getConnection()
-    .getRepository(Project)
-    .find({
-      where: { uuid: id },
-    });
-  return res.status(OK).json({ projects });
+  const { user } = req;
+  if (user) {
+    console.log(user.uuid);
+    const projects = await getConnection()
+      .getRepository(Project)
+      .find({
+        where: { uuid: user.uuid },
+      });
+    return res.status(OK).json({ projects });
+  } else {
+    res.status(FORBIDDEN).end();
+  }
 };
 
 /******************************************************************************
@@ -51,51 +45,60 @@ export const one = async (
   req: Request,
   res: Response
 ): Promise<Response | void> => {
-  const { id } = req.params as ParamsDictionary;
-  const project = await getConnection().getRepository(Project).findOne(id);
-  const toDoTasks = await getConnection()
-    .getRepository(Task)
-    .find({
-      where: { project: { id }, taskStatus: "todo" },
-      order: { position: "ASC" },
-    });
-  const doingTasks = await getConnection()
-    .getRepository(Task)
-    .find({
-      where: { project: { id }, taskStatus: "doing" },
-      order: { position: "ASC" },
-    });
-  const doneTasks = await getConnection()
-    .getRepository(Task)
-    .find({
-      where: { project: { id }, taskStatus: "done" },
-      order: { position: "ASC" },
-    });
-  const materialItems = await getConnection()
-    .getRepository(Item)
-    .find({ where: { project: { id }, category: "material" } });
-  const laborItems = await getConnection()
-    .getRepository(Item)
-    .find({ where: { project: { id }, category: "labor" } });
-  const otherItems = await getConnection()
-    .getRepository(Item)
-    .find({ where: { project: { id }, category: "other" } });
-  if (!project) {
-    res.status(NOT_FOUND);
-    res.end();
-    return;
-  }
+  const { user } = req;
+  if (user) {
+    const { id } = req.params as ParamsDictionary;
+    const project = await getConnection().getRepository(Project).findOne(id);
+    const toDoTasks = await getConnection()
+      .getRepository(Task)
+      .find({
+        where: { project: { id }, taskStatus: "todo" },
+        order: { position: "ASC" },
+      });
+    const doingTasks = await getConnection()
+      .getRepository(Task)
+      .find({
+        where: { project: { id }, taskStatus: "doing" },
+        order: { position: "ASC" },
+      });
+    const doneTasks = await getConnection()
+      .getRepository(Task)
+      .find({
+        where: { project: { id }, taskStatus: "done" },
+        order: { position: "ASC" },
+      });
+    const materialItems = await getConnection()
+      .getRepository(Item)
+      .find({ where: { project: { id }, category: "material" } });
+    const laborItems = await getConnection()
+      .getRepository(Item)
+      .find({ where: { project: { id }, category: "labor" } });
+    const otherItems = await getConnection()
+      .getRepository(Item)
+      .find({ where: { project: { id }, category: "other" } });
+    if (!project) {
+      res.status(NOT_FOUND);
+      res.end();
+      return;
+    }
 
-  return res.status(OK).json({
-    currentProject: {
-      projectName: project.projectName,
-      startDate: project.startDate,
-      endDate: project.endDate,
-      id: project.id,
-      items: { material: materialItems, labor: laborItems, other: otherItems },
-      tasks: { todo: toDoTasks, doing: doingTasks, done: doneTasks },
-    },
-  });
+    return res.status(OK).json({
+      currentProject: {
+        projectName: project.projectName,
+        startDate: project.startDate,
+        endDate: project.endDate,
+        id: project.id,
+        items: {
+          material: materialItems,
+          labor: laborItems,
+          other: otherItems,
+        },
+        tasks: { todo: toDoTasks, doing: doingTasks, done: doneTasks },
+      },
+    });
+  } else {
+    res.status(FORBIDDEN).end();
+  }
 };
 
 /******************************************************************************
@@ -106,22 +109,27 @@ export const add = async (
   req: Request,
   res: Response
 ): Promise<Response | void> => {
-  const { project: input } = req.body;
-  const project = new Project();
-  project.projectName = input.projectName;
-  project.startDate = input.startDate;
-  project.endDate = input.endDate;
-  project.uuid = input.uuid;
-  const errors = await validate(project);
+  const { user } = req;
+  if (user) {
+    const { project: input } = req.body;
+    const project = new Project();
+    project.projectName = input.projectName;
+    project.startDate = input.startDate;
+    project.endDate = input.endDate;
+    project.uuid = input.uuid;
+    const errors = await validate(project);
 
-  if (errors.length > 0) {
-    logger.error(errors);
-    res.status(BAD_REQUEST).json({ error: errors }).end();
-    return;
+    if (errors.length > 0) {
+      logger.error(errors);
+      res.status(BAD_REQUEST).json({ error: errors }).end();
+      return;
+    }
+
+    const data = await getConnection().getRepository(Project).save(project);
+    return res.status(CREATED).json({ project: data });
+  } else {
+    res.status(FORBIDDEN).end();
   }
-
-  const data = await getConnection().getRepository(Project).save(project);
-  return res.status(CREATED).json({ project: data });
 };
 
 /******************************************************************************
@@ -132,19 +140,24 @@ export const update = async (
   req: Request,
   res: Response
 ): Promise<Response | void> => {
-  const { project } = req.body;
-  if (!project && !project.id) {
-    res
-      .status(BAD_REQUEST)
-      .json({
-        error: paramMissingError,
-      })
-      .end();
-    return;
+  const { user } = req;
+  if (user) {
+    const { project } = req.body;
+    if (!project && !project.id) {
+      res
+        .status(BAD_REQUEST)
+        .json({
+          error: paramMissingError,
+        })
+        .end();
+      return;
+    }
+    // add validation and only set provided fields
+    const data = await getConnection().getRepository(Project).save(project);
+    return res.status(OK).json({ project: data });
+  } else {
+    res.status(FORBIDDEN).end();
   }
-  // add validation and only set provided fields
-  const data = await getConnection().getRepository(Project).save(project);
-  return res.status(OK).json({ project: data });
 };
 
 /******************************************************************************
@@ -155,16 +168,21 @@ export const remove = async (
   req: Request,
   res: Response
 ): Promise<Response | void> => {
-  const repository = await getConnection().getRepository(Project);
-  const { id } = req.params as ParamsDictionary;
-  const project = await repository.findOne(id);
-  if (!project) {
-    res.status(BAD_REQUEST);
-    res.end();
-    return;
+  const { user } = req;
+  if (user) {
+    const repository = await getConnection().getRepository(Project);
+    const { id } = req.params as ParamsDictionary;
+    const project = await repository.findOne(id);
+    if (!project) {
+      res.status(BAD_REQUEST);
+      res.end();
+      return;
+    }
+    await repository.remove([project]);
+    return res.status(OK).end();
+  } else {
+    res.status(FORBIDDEN).end();
   }
-  await repository.remove([project]);
-  return res.status(OK).end();
 };
 
 /******************************************************************************
@@ -177,5 +195,4 @@ export default {
   add,
   update,
   remove,
-  getmine,
 };
